Added the two files
main_script.py is the main python file with the code for: (i) gathering the data, (ii) processing any signals, (iii) messaging any signals. You will need to input and change some of the information in this file to match your data accordingly, such as your Twilio account number. 

webserver.py is used so that it can be linked to UptimeRobot such that the repl script does not stop running from inactivity. It keeps the script running perpetually. Steps on how to monitor the script on UptimeRobot so it runs forever are written as comments in this python file.